build cli
