--- cli/.gitignore
@@ -1,4 +1,0 @@
-dist
-Buffered.link.egg-info
-build
-__pycache__
