+++ cli/Buffered.link.egg-info/requires.txt
@@ -1,0 +1,2 @@
+requests
+rich
