+++ cli/setup.py
@@ -6,13 +6,14 @@ with open("README.md", "r") as desc:
 
 setuptools.setup(
     name="Buffered.link",
-    version="1.0.1",
+    version="1.0.3",
     author="Josias Aurel",
     description="You text buffer CLI",
     long_description=long_description,
     long_description_content_type="text/markdown",
     packages=setuptools.find_packages(),
     python_requires=">=3.6",
-    py_modules=["buffered-link"],
-    install_requires=["requests", "rich"]
+    py_modules=["bfdl"],
+    install_requires=["requests", "rich"],
+    author_email="josias@josiasw.dev"
 )
